+++ tapo-py/src/handlers/child_devices.rs
@@ -1,7 +1,9 @@
+mod t100_handler;
 mod t110_handler;
 mod t300_handler;
 mod t31x_handler;
 
+pub use t100_handler::*;
 pub use t110_handler::*;
 pub use t300_handler::*;
 pub use t31x_handler::*;