+++ tapo-py/src/lib.rs
@@ -9,8 +9,8 @@ use api_client::PyApiClient;
 use handlers::{
     PyColorLightHandler, PyColorLightSetDeviceInfoParams, PyEnergyDataInterval,
     PyGenericDeviceHandler, PyHubHandler, PyLightHandler, PyPlugEnergyMonitoringHandler,
-    PyPlugHandler, PyT110Handler, PyT300Handler, PyT31XHandler, TriggerLogsT110Result,
-    TriggerLogsT300Result,
+    PyPlugHandler, PyT110Handler, PyT300Handler, PyT31XHandler, TriggerLogsT100Result,
+    TriggerLogsT110Result, TriggerLogsT300Result,
 };
 use tapo::requests::Color;
 use tapo::responses::{
@@ -20,7 +20,7 @@ use tapo::responses::{
     DeviceInfoLightResult, DeviceInfoPlugEnergyMonitoringResult, DeviceInfoPlugResult,
     DeviceUsageEnergyMonitoringResult, DeviceUsageResult, EnergyDataResult, EnergyUsageResult,
     KE100Result, OvercurrentStatus, OverheatStatus, PlugState, PowerProtectionStatus, S200BResult,
-    Status, T100Result, T110Log, T110Result, T300Log, T300Result, T31XResult,
+    Status, T100Log, T100Result, T110Log, T110Result, T300Log, T300Result, T31XResult,
     TemperatureHumidityRecord, TemperatureHumidityRecords, TemperatureUnit, TemperatureUnitKE100,
     UsageByPeriodResult, WaterLeakStatus,
 };
@@ -81,12 +81,14 @@ fn tapo_py(py: Python, module: &Bound<'_, PyModule>) -> PyResult<()> {
 
     // responses: hub devices
     responses.add_class::<Status>()?;
+    responses.add_class::<T100Log>()?;
     responses.add_class::<T110Log>()?;
     responses.add_class::<T300Log>()?;
     responses.add_class::<TemperatureHumidityRecord>()?;
     responses.add_class::<TemperatureHumidityRecords>()?;
     responses.add_class::<TemperatureUnit>()?;
     responses.add_class::<TemperatureUnitKE100>()?;
+    responses.add_class::<TriggerLogsT100Result>()?;
     responses.add_class::<TriggerLogsT110Result>()?;
     responses.add_class::<TriggerLogsT300Result>()?;
     responses.add_class::<WaterLeakStatus>()?;